Treat setting to null or undefined as unsetting
In the session data, if you set an attribute to null or undefined,
and then read that variable again, it would return the string
"null" or "undefined" instead of the values null or undefined.
This fixes part of the problem. If you set the value of an attribute
to null or undefined, then it will unset the attribute as per the
HttpServlet 3.1 javadocs instead of setting it to a string. This
is still not quite working correctly, because if you unset a
variable and then test its value, it always returns null instead of
undefined.